tourneyhub/requirements.txt
asgiref==3.8.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.0.6
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
openapi-codec==1.3.2
packaging==24.1
platformdirs==4.2.2
simplejson==3.19.2
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
virtualenv==20.26.2
